@@ -1,11 +1,9 @@
 [project]
 name = "tap-stackoverflow-sampledata"
-version = "0.1.3"
+version = "0.2.0"
 description = "`tap-stackoverflow-sampledata` is a Singer tap for stackoverflow-sampledata, built with the Meltano SDK for Singer Taps."
-authors = [{ name = "Dan Norman", email = "buzzcutnorman@gmail.com" }]
-requires-python = ">=3.8"
 readme = "README.md"
-license = "MIT"
+authors = [{ name = "Dan Norman", email = "buzzcutnorman@gmail.com" }]
 keywords = [
     "ELT",
     "stackoverflow-sampledata",
@@ -13,43 +11,49 @@ keywords = [
 classifiers = [
     "Intended Audience :: Developers",
     "Operating System :: OS Independent",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
 ]
+license = "MIT"
+license-files = [ "LICENSE.md" ]
+requires-python = ">=3.9"
 dependencies = [
-    "importlib-resources==6.4.* ; python_version < '3.9'",
-    "singer-sdk[]~=0.39.0",
-    "lxml>=5.2.2,<6",
+    "singer-sdk[msgspec]~=0.46.4",
+    "lxml>=5.4,<6",
 ]
 
 [project.optional-dependencies]
-s3 = ["fs-s3fs~=1.1.1"]
+s3 = [
+    "fs-s3fs~=1.1.1",
+]
 
 [project.scripts]
+# CLI declaration
 tap-stackoverflow-sampledata = "tap_stackoverflow_sampledata.tap:TapStackOverflowSampleData.cli"
 
 [dependency-groups]
 dev = [
+    { include-group = "test" },
+]
+test = [
     "pytest>=8",
-    "singer-sdk[testing]~=0.39.0",
+    "singer-sdk[testing]",
 ]
 
-[build-system]
-requires = [
-    "hatchling>=1,<2",
+[tool.pytest.ini_options]
+addopts = [
+    "--durations=10",
 ]
-build-backend = "hatchling.build"
 
 [tool.mypy]
-python_version = "3.9"
 warn_unused_configs = true
 
 [tool.ruff]
 src = ["tap_stackoverflow_sampledata"]
-target-version = "py38"
+target-version = "py39"
 
 [tool.ruff.lint]
 ignore = [
@@ -63,8 +67,11 @@ select = ["ALL"]
 [tool.ruff.lint.flake8-annotations]
 allow-star-arg-any = true
 
-[tool.ruff.lint.isort]
-known-first-party = ["tap_stackoverflow_sampledata"]
-
 [tool.ruff.lint.pydocstyle]
 convention = "google"
+
+[build-system]
+requires = [
+    "hatchling>=1,<2",
+]
+build-backend = "hatchling.build"
