@@ -3,12 +3,12 @@
 from __future__ import annotations
 
 import os
-from typing import TYPE_CHECKING, Iterable
+import typing as t
 
 from lxml import etree
 from singer_sdk.streams import Stream
 
-if TYPE_CHECKING:
+if t.TYPE_CHECKING:
     from singer_sdk.helpers.types import Context
 
 class NonExistentDataDirectoryError(Exception):
@@ -39,7 +39,7 @@ class StackOverflowSampleDataStream(Stream):
     def get_records(
         self,
         context: Context | None,  # noqa: ARG002
-    ) -> Iterable[dict]:
+    ) -> t.Iterable[dict]:
         """Return a generator of row-type dictionary objects.
 
         Args:
